update checkSharedBytes clang format
