+++ include/gauge_fix_ovr_hit_devf.cuh
@@ -82,13 +82,13 @@ namespace quda {
       if (mu < gauge_dir || (mu >= 4 && mu < (gauge_dir + 4))) {
         //Retrieve the four SU(2) parameters...
         // a0
-        atomic_add_local(elems + tid, (link(p,p)).x + (link(q,q)).x); //a0
+        atomic_add_local(elems + tid, (link(p, p)).x + (link(q, q)).x); // a0
         // a1
-        atomic_add_local(elems + tid + blockSize, (link(p,q).y + link(q,p).y) * asq); //a1
+        atomic_add_local(elems + tid + blockSize, (link(p, q).y + link(q, p).y) * asq); // a1
         // a2
-        atomic_add_local(elems + tid + blockSize * 2, (link(p,q).x - link(q,p).x) * asq); //a2
+        atomic_add_local(elems + tid + blockSize * 2, (link(p, q).x - link(q, p).x) * asq); // a2
         // a3
-        atomic_add_local(elems + tid + blockSize * 3, (link(p,p).y - link(q,q).y) * asq); //a3
+        atomic_add_local(elems + tid + blockSize * 3, (link(p, p).y - link(q, q).y) * asq); // a3
       } //FLOP per lattice site = gauge_dir * 2 * (4 + 7) = gauge_dir * 22
 
       cache.sync();
@@ -366,13 +366,13 @@ namespace quda {
       if (mu < gauge_dir) {
         //Retrieve the four SU(2) parameters...
         // a0
-        atomic_add_local(elems + tid, (link1(p,p)).x + (link1(q,q)).x + (link(p,p)).x + (link(q,q)).x); //a0
+        atomic_add_local(elems + tid, (link1(p, p)).x + (link1(q, q)).x + (link(p, p)).x + (link(q, q)).x); // a0
         // a1
-        atomic_add_local(elems + tid + blockSize, (link1(p,q).y + link1(q,p).y) - (link(p,q).y + link(q,p).y)); //a1
+        atomic_add_local(elems + tid + blockSize, (link1(p, q).y + link1(q, p).y) - (link(p, q).y + link(q, p).y)); // a1
         // a2
-        atomic_add_local(elems + tid + blockSize * 2, (link1(p,q).x - link1(q,p).x) - (link(p,q).x - link(q,p).x)); //a2
+        atomic_add_local(elems + tid + blockSize * 2, (link1(p, q).x - link1(q, p).x) - (link(p, q).x - link(q, p).x)); // a2
         // a3
-        atomic_add_local(elems + tid + blockSize * 3, (link1(p,p).y - link1(q,q).y) - (link(p,p).y - link(q,q).y)); //a3
+        atomic_add_local(elems + tid + blockSize * 3, (link1(p, p).y - link1(q, q).y) - (link(p, p).y - link(q, q).y)); // a3
       }
       cache.sync();
       if (mu == 0) {
