+++ include/kernels/restrictor.cuh
@@ -137,8 +137,8 @@ namespace quda {
     using BlockReduce_t = typename RestrictorParams<Arg>::BlockReduce_t;
     const Arg &arg;
     using typename KernelOps<BlockReduce_t>::KernelOpsT;
-    template <typename ...Ops>
-    constexpr Restrictor(const Arg &arg, const Ops &...ops) : KernelOpsT(ops...), arg(arg) {}
+    template <typename... Ops>
+    constexpr Restrictor(const Arg &arg, const Ops &...ops) : KernelOpsT(ops...), arg(arg) { }
     static constexpr const char *filename() { return KERNEL_FILE; }
 
     __device__ __host__ inline void operator()(dim3 block, dim3 thread)
@@ -200,5 +200,4 @@ namespace quda {
       }
     }
   };
-
 }
