--- include/targets/cuda/shared_memory_cache_helper.h
@@ -1,1 +1,0 @@
-#include "../generic/shared_memory_cache_helper.h"
