--- include/targets/cuda/thread_local_cache.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#if 1
-#include "../generic/thread_local_cache.h"
-#else
-#include "../generic/thread_local_cache_noshared.h"
-#endif
