+++ lib/dslash_ndeg_twisted_clover.cu
@@ -24,10 +24,10 @@ namespace quda
 
       unsigned int sharedBytesPerThread() const
       {
-	if (arg.kernel_type == INTERIOR_KERNEL) {
-	  return 2 * in.Ncolor() * 4 * sizeof(typename mapper<typename Arg::Float>::type);
-	}
-	return 0;
+        if (arg.kernel_type == INTERIOR_KERNEL) {
+          return 2 * in.Ncolor() * 4 * sizeof(typename mapper<typename Arg::Float>::type);
+        }
+        return 0;
       }
 
     public:
